refactor(#310): ProcessTapEngineをファイル分割しCoreAudio呼び出しをAudioTapGateway経由に
400行制限の規約に沿い ProcessTapEngine.swift を6ファイルへ分割
(LifecycleState / CATapDescription+ProcessTap / ProcessSubtree /
TapSampleRate / UnsafeMutableAudioBufferListPointer+Deinterleave /
CoreAudioTapGateway)。

「時間・空間・資源・乱数・操作は抽象化されるべき」という方針に従い、
ProcessTapEngine.init? と LifecycleState.rolledBack() が直接呼んでいた
CoreAudio C API (AudioHardwareCreateProcessTap 等9呼び出し) を
Domain.AudioTapGateway プロトコル越しに注入する形へ変更。ProcessGateway/
DarwinGateway と同じ「純粋ロジック + 薄いOS境界シェル」パターンをこの
モジュールにも適用し、実機オーディオなしで init? の全分岐(成功/各失敗/
ロールバック順序)とIOProcコールバック本体をユニットテスト化した。

カバレッジ: AudioTapDataSourceモジュール 48.21%→75.44%(line)。
CoreAudioTapGateway.swift(実CoreAudio呼び出しの薄いシェル、DarwinGateway
と同じ0%が想定内)を除く全ファイルが87.5%〜100%、ProcessTapEngine.swift
自体は100%。
